enhanced_summary/requirements.txt
google-generativeai
pandas
python-dotenv
streamlit
pathlib